+++ CHANGELOG.md
@@ -36,8 +36,7 @@ To Document:
 
 TO DO:
 * IN PROGRESS: Factor out RunResult
-  * NEXT: Convert beforeEach/afterEach from TestCaseResult[] to RunResult[]
-  * THEN: Move createPass() etc. factories to RunResult
+  * NEXT: Convert beforeAll/afterAll from TestCaseResult[] to RunResult[]
   * Convert TestResult.suite() to TestSuiteResult.create() and TestResult.testCase() to TestCaseResult.create()
   * Revise TestSuiteResult.create() to match signature of other result factories
   * THEN: Modify TestSuite to return RunResult when running test function rather than TestCaseResult